Restructure as proper Python package
- Create src/tpane/ package directory with proper structure
- Move tpane.py to src/tpane/__main__.py for python -m support
- Reorganize core/ and parsers/ as subpackages under tpane/
- Update test imports to use new package structure
- Resolves Copilot feedback about entry point configuration

This eliminates ambiguity between module and package structure,
follows Python packaging best practices, and enables both:
- python -m tpane (module execution)
- pip install tpane (with tpane command)
