+++ src/tpane/__init__.py
@@ -4,7 +4,7 @@ tpane - Reference implementation of TOPA (Test Output Protocol for AI)
 A standardized test output format designed for LLM consumption.
 """
 
-from .tpane import main
+from .__main__ import main
 
 __version__ = "0.1.0"
 __all__ = ["main"]
